+++ openfisca_us/variables/irs/outputs.py
@@ -375,6 +375,120 @@ class taxbc(Variable):
     definition_period = YEAR
     documentation = """Regular tax on regular taxable income before credits"""
 
+    def formula(tax_unit, period, parameters):
+        capital_gains = parameters(period).irs.capital_gains.brackets
+        mars = tax_unit("mars", period)
+        dwks1 = tax_unit("c04800", period)
+
+        dwks16 = min_(capital_gains.thresholds["1"][mars], dwks1)
+        dwks17 = min_(tax_unit("dwks14", period), dwks16)
+
+        dwks20 = dwks16 - dwks17
+        lowest_rate_tax = capital_gains.rates["1"] * dwks20
+        # Break in worksheet lines
+        dwks13 = tax_unit("dwks13", period)
+        dwks21 = min_(dwks1, dwks13)
+        dwks22 = dwks20
+        dwks23 = max_(0, dwks21 - dwks22)
+        dwks25 = min_(capital_gains.thresholds["2"][mars], dwks1)
+        dwks19 = tax_unit("dwks19", period)
+        dwks26 = min_(dwks19, dwks20)
+        dwks27 = max_(0, dwks25 - dwks26)
+        dwks28 = min_(dwks23, dwks27)
+        dwks29 = capital_gains.rates["2"] * dwks28
+        dwks30 = dwks22 + dwks28
+        dwks31 = dwks21 - dwks30
+        dwks32 = capital_gains.rates["3"] * dwks31
+        # Break in worksheet lines
+        dwks33 = min_(
+            tax_unit("dwks9", period), tax_unit("filer_e24515", period)
+        )
+        dwks10 = tax_unit("dwks10", period)
+        dwks34 = dwks10 + dwks19
+        dwks36 = max_(0, dwks34 - dwks1)
+        dwks37 = max_(0, dwks33 - dwks36)
+        dwks38 = 0.25 * dwks37
+        # Break in worksheet lines
+        dwks39 = dwks19 + dwks20 + dwks28 + dwks31 + dwks37
+        dwks40 = dwks1 - dwks39
+        dwks41 = 0.28 * dwks40
+
+        # SchXYZ call in Tax-Calculator
+
+        # Separate non-negative taxable income into two non-negative components,
+        # doing this in a way so that the components add up to taxable income
+        # define pass-through income eligible for PT schedule
+        individual_income = parameters(period).irs.income
+        e26270 = tax_unit("filer_e26270", period)
+        e00900 = tax_unit("filer_e00900", period)
+
+        # Determine pass-through and non-pass-through income
+        pt_active_gross = e00900 + e26270
+        pt_active = pt_active_gross
+        pt_active = min_(pt_active, e00900 + e26270)
+        pt_taxinc = max_(0, pt_active)
+        taxable_income = dwks19
+
+        pt_taxinc = min_(pt_taxinc, taxable_income)
+        reg_taxinc = max_(0, taxable_income - pt_taxinc)
+        pt_tbase = reg_taxinc
+
+        mars = tax_unit("mars", period)
+
+        # Initialise regular and pass-through income tax to zero
+        reg_tax = 0
+        pt_tax = 0
+        last_reg_threshold = 0
+        last_pt_threshold = 0
+        for i in range(1, 7):
+            # Calculate rate applied to regular income up to the current
+            # threshold (on income above the last threshold)
+            reg_threshold = individual_income.bracket.thresholds[str(i)][mars]
+            reg_tax += individual_income.bracket.rates[
+                str(i)
+            ] * amount_between(reg_taxinc, last_reg_threshold, reg_threshold)
+            last_reg_threshold = reg_threshold
+
+            # Calculate rate applied to pass-through income on in the same
+            # way, but as treated as if stacked on top of regular income
+            # (which is not taxed again)
+            pt_threshold = max(
+                individual_income.pass_through.bracket.thresholds[str(i)][mars]
+                - pt_tbase,
+                0,
+            )
+            pt_tax += individual_income.pass_through.bracket.rates[
+                str(i)
+            ] * amount_between(pt_taxinc, last_pt_threshold, pt_threshold)
+            last_pt_threshold = pt_threshold
+
+        # Calculate regular and pass-through tax above the last threshold
+        reg_tax += individual_income.bracket.rates["7"] * max_(
+            reg_taxinc - last_reg_threshold, 0
+        )
+        pt_tax += individual_income.pass_through.bracket.rates["7"] * max_(
+            pt_taxinc - last_pt_threshold, 0
+        )
+
+        dwks42 = reg_tax + pt_tax
+        dwks43 = sum(
+            [
+                dwks29,
+                dwks32,
+                dwks38,
+                dwks41,
+                dwks42,
+                lowest_rate_tax,
+            ]
+        )
+        c05200 = tax_unit("c05200", period)
+        dwks44 = c05200
+        dwks45 = min_(dwks43, dwks44)
+
+        hasqdivltcg = tax_unit("hasqdivltcg", period)
+
+        return where(hasqdivltcg, dwks45, c05200)
+
 
 class c00100(Variable):
     value_type = float
@@ -1323,40 +1437,126 @@ class charity_credit(Variable):
     documentation = """Credit for charitable giving"""
 
 
+class dwks6(Variable):
+    value_type = float
+    entity = TaxUnit
+    label = "DWKS6"
+    unit = "currency-USD"
+    definition_period = YEAR
+
+    def formula(tax_unit, period, parameters):
+        dwks2 = tax_unit("filer_e00650", period)
+        dwks3 = tax_unit("filer_e58990", period)
+        # dwks4 always assumed to be zero
+        dwks5 = max_(0, dwks3)
+        return max_(0, dwks2 - dwks5)
+
+
+class dwks9(Variable):
+    value_type = float
+    entity = TaxUnit
+    label = "DWKS9"
+    unit = "currency-USD"
+    definition_period = YEAR
+
+    def formula(tax_unit, period, parameters):
+        p23250 = tax_unit("filer_p23250", period)
+        c23650 = tax_unit("c23650", period)
+        dwks7 = min_(p23250, c23650)  # SchD lines 15 and 16, respectively
+        # dwks8 = min(dwks3, dwks4)
+        # dwks9 = max(0., dwks7 - dwks8)
+        # BELOW TWO STATEMENTS ARE UNCLEAR IN LIGHT OF dwks9=... COMMENT
+        e01100 = tax_unit("filer_e01100", period)
+        c24510 = where(e01100 > 0, e01100, max_(0, dwks7) + e01100)
+        return max_(0, c24510 - min_(0, tax_unit("filer_e58990", period)))
+
+
 class dwks10(Variable):
     value_type = float
     entity = TaxUnit
     definition_period = YEAR
+    label = "DWKS10"
     documentation = (
         """search taxcalc/calcfunctions.py for how calculated and used"""
     )
+
+    def formula(tax_unit, period, parameters):
+        dwks10_if_gains = add(tax_unit, period, "dwks6", "dwks9")
+        dwks10_if_no_gains = (
+            max_(
+                0,
+                min_(
+                    tax_unit("filer_p23250", period),
+                    tax_unit("c23650", period),
+                ),
+            )
+            + tax_unit("filer_e01100", period)
+        )
+        return where(
+            tax_unit("hasqdivltcg", period),
+            dwks10_if_gains,
+            dwks10_if_no_gains,
+        )
 
 
 class dwks13(Variable):
     value_type = float
     entity = TaxUnit
     definition_period = YEAR
+    label = "DWKS13"
+    unit = "currency-USD"
     documentation = (
         """search taxcalc/calcfunctions.py for how calculated and used"""
     )
+
+    def formula(tax_unit, period, parameters):
+        dwks1 = tax_unit("c04800", period)
+        e24515 = tax_unit("filer_e24515", period)
+        dwks11 = e24515 + tax_unit(
+            "filer_e24518", period
+        )  # Sch D lines 18 and 19, respectively
+        dwks9 = tax_unit("dwks9", period)
+        dwks12 = min_(dwks9, dwks11)
+        dwks10 = tax_unit("dwks10", period)
+        return (dwks10 - dwks12) * tax_unit("hasqdivltcg", period)
 
 
 class dwks14(Variable):
     value_type = float
     entity = TaxUnit
     definition_period = YEAR
+    label = "DWKS14"
+    unit = "currency-USD"
     documentation = (
         """search taxcalc/calcfunctions.py for how calculated and used"""
     )
+
+    def formula(tax_unit, period, parameters):
+        dwks1 = tax_unit("c04800", period)
+        dwks13 = tax_unit("dwks13", period)
+        return max_(0, dwks1 - dwks13) * tax_unit("hasqdivltcg", period)
 
 
 class dwks19(Variable):
     value_type = float
     entity = TaxUnit
     definition_period = YEAR
+    label = "DWKS14"
+    unit = "currency-USD"
     documentation = (
         """search taxcalc/calcfunctions.py for how calculated and used"""
     )
+
+    def formula(tax_unit, period, parameters):
+        dwks14 = tax_unit("dwks14", period)
+        capital_gains = parameters(period).irs.capital_gains.brackets
+        mars = tax_unit("mars", period)
+        dwks1 = tax_unit("c04800", period)
+        dwks16 = min_(capital_gains.thresholds["1"][mars], dwks1)
+        dwks17 = min_(dwks14, dwks16)
+        dwks10 = tax_unit("dwks10", period)
+        dwks18 = max_(0, dwks1 - dwks10)
+        return max_(dwks17, dwks18) * tax_unit("hasqdivltcg", period)
 
 
 class fstax(Variable):