+++ CHANGELOG.md
@@ -26,6 +26,14 @@
 * Added Alternative Minimum Tax (AMT) income and liability logic
 * Added development tools for auto-generating unit tests for Tax-Calculator functions
 
-### 0.7.0
+## 0.7.0
+
+* Added Gains Tax (capital gains treatment) logic and parameters
+
+### 0.8.0
+
+* Added minimum benefit logic for SNAP
+
+### 0.9.0
 
 * Added Social Security taxation logic